# Try sync.Map as cache for index