--- a/datalog.go
+++ b/datalog.go
@@ -74,7 +74,7 @@ func (d *Datalog) RebuildIndex() error {
 		size := int(binary.BigEndian.Uint32(buf[:intSize]))
 		var score Score
 		copy(score[:], buf[intSize:])
-		err = d.index.Write(score, Addr{pos: pos, size: size})
+		err = d.index.Store(score, Addr{pos: pos, size: size})
 		if err != nil {
 			break
 		}
@@ -86,7 +86,7 @@ func (d *Datalog) RebuildIndex() error {
 
 // Read reads data for a given position and length
 func (d *Datalog) Read(score Score) ([]byte, error) {
-	a, ok := d.index.Read(score)
+	a, ok := d.index.Load(score)
 	if !ok {
 		err := fmt.Errorf("Unknown score %s", score)
 		return nil, err
@@ -105,7 +105,7 @@ func (d *Datalog) Read(score Score) ([]byte, error) {
 // Write writes given data into datalog and returns it's position and length
 func (d *Datalog) Write(data []byte) (Score, error) {
 	score := MakeScore(data)
-	if _, ok := d.index.Read(score); ok {
+	if _, ok := d.index.Load(score); ok {
 		return score, nil
 	}
 	size := len(data) + scoreSize
@@ -120,7 +120,7 @@ func (d *Datalog) Write(data []byte) (Score, error) {
 	pos := int(d.cur) + intSize
 	size = n - intSize
 	d.cur += n
-	err = d.index.Write(score, Addr{pos: pos, size: size})
+	err = d.index.Store(score, Addr{pos: pos, size: size})
 	if err != nil {
 		return Score{}, err
 	}
@@ -130,7 +130,7 @@ func (d *Datalog) Write(data []byte) (Score, error) {
 // Close closes the datalog
 func (d *Datalog) Close() error {
 	d.indexRW.Close()
-	d.index = &Index{}
+	d.index = new(Index)
 	d.cur = 0
 	return d.rwc.Close()
 }