Just to see how it'd hold vs "normal" map