@eiri 2019-04-04
## Bench

```
benchmark                   old ns/op      new ns/op      delta
BenchmarkRebuildIndex-4     3058643446     3455678309     +12.98%
BenchmarkOpen-4             81379017       252409192      +210.16%
BenchmarkWrite-4            25165          28135          +11.80%
BenchmarkRead-4             3955           3891           -1.62%

benchmark                   old allocs     new allocs     delta
BenchmarkRebuildIndex-4     245708         1424992        +479.95%
BenchmarkOpen-4             11000          1190295        +10720.86%
BenchmarkWrite-4            2              7              +250.00%
BenchmarkRead-4             1              1              +0.00%

benchmark                   old bytes     new bytes     delta
BenchmarkRebuildIndex-4     48646032      55510112      +14.11%
BenchmarkOpen-4             46757667      53621486      +14.68%
BenchmarkWrite-4            1291          1409          +9.14%
BenchmarkRead-4             16            16            +0.00%
```