@eiri 2019-04-04
## Summary

Whoa, big hit on opening of datalog, i.e. index load! That `store` penalty quite noticeable, so only useful switch for concurrent access, with sharded datalogs and single index or something.